Updating EntityFramework version to 6.0.0 and updating XmlGenerator project accordingly.

+++ Tools/XmlGenerator/StaticData/DgmTypeAttributes.cs
@@ -15,7 +15,7 @@ namespace EVEMon.XmlGenerator.StaticData
         public Nullable<Int64> ValueInt64 { get; set; }
 
         [XmlElement("valueFloat")]
-        public Nullable<Single> ValueFloat { get; set; }
+        public Nullable<Double> ValueFloat { get; set; }
 
         /// <summary>
         /// Returns the value as an integer.
